refactor: apply Strategy Pattern to WidgetRenderer and centralize default sizes in configuration
